poc(parsing): Leverage TOML's array parser to extract arguments
This is a bit of a hack, but it saves me from writing a whole custom parser.
It might be better to switch over to named arguments if TOML parses as key-value pairs.
Further fixes are required for enforcing positive values and unsigned int types.

Fundamentally this will all have to be rewritten, but it's a good experiment.
The core point is that I've built enough familiarity with Serde to guide the rewrite.